+++ packages/filter/src/components/panel/FilterPanelBar.tsx
@@ -1,9 +1,13 @@
-import { Button } from '@equinor/fusion-react-button';
+import { Button, Icon } from '@equinor/eds-core-react';
+import { chevron_up, chevron_down, refresh } from '@equinor/eds-icons';
+
 import { clsx, createStyles, makeStyles } from '@equinor/fusion-react-styles';
 import { FilterFn } from '../../types';
 import { SearchFilter } from '../filter';
 import { ClearFilterButton } from '../misc';
 import { useFilterPanelContext } from './FilterPanelProvider';
+
+Icon.add({ chevron_down, chevron_up, refresh });
 
 const useStyles = makeStyles(
   (theme) =>
@@ -40,18 +44,24 @@ export const FilterPanelBar = <TData,>(props: FilterPanelBarProps<TData>): JSX.E
   const { className, searchFn, ...args } = props;
   const { showFilters, setShowFilters } = useFilterPanelContext();
   const styles = useStyles();
+
   return (
     <div {...args} className={clsx(className, styles.root)}>
-      <SearchFilter filterKey="global" label="Search all" className={styles.searchInput} dense filterFn={searchFn} />
+      <SearchFilter filterKey="global" label="Search all" className={styles.searchInput} filterFn={searchFn} />
       <div className={styles.actions}>
-        <ClearFilterButton className={styles.resetBtn} label="Reset Filters" icon="refresh" variant="ghost" />
+        <ClearFilterButton className={styles.resetBtn} variant="ghost">
+          Reset Filters
+          <Icon name="refresh" />
+        </ClearFilterButton>
         <Button
           icon={showFilters ? 'chevron_up' : 'chevron_down'}
-          label={showFilters ? 'Hide filters' : 'Show filters'}
           onClick={() => setShowFilters(!showFilters)}
           variant="outlined"
           trailingIcon
-        />
+        >
+          {showFilters ? 'Hide filters' : 'Show filters'}
+          <Icon name={showFilters ? 'chevron_up' : 'chevron_down'} />
+        </Button>
       </div>
     </div>
   );