+++ packages/person/src/PersonAvatar.tsx
@@ -4,7 +4,12 @@ import HTMLPersonAvatarCustomElement, { tag, AvatarSize, AvatarData } from '@equ
 import extractProps from './extract-props';
 
 type ElementProps = PropsWithChildren<
-  Partial<Pick<HTMLPersonAvatarCustomElement, 'azureId' | 'upn' | 'dataSource' | 'size' | 'clickable' | 'disabled'>>
+  Partial<
+    Pick<
+      HTMLPersonAvatarCustomElement,
+      'azureId' | 'upn' | 'dataSource' | 'size' | 'clickable' | 'disabled' | 'trigger'
+    >
+  >
 >;
 
 export type PersonAvatarProps = ComponentProps<HTMLPersonAvatarCustomElement, ElementProps>;
@@ -25,7 +30,7 @@ export const PersonAvatar = ({ children, ...props }: PropsWithChildren<PersonAva
         avatarRef.current[name] = value;
       }
     }
-  }, []);
+  }, [props]);
 
   return <PersonAvatarComponent ref={avatarRef}>{children}</PersonAvatarComponent>;
 };