+++ packages/progress-indicator/src/index.ts
@@ -1,3 +1,0 @@
-export * from './star';
-export * from './circular';
-export * from './dots';