+++ packages/side-sheet/src/components/Content.tsx
@@ -4,6 +4,7 @@ import styled from 'styled-components';
 const StyledContentContainer = styled.div`
   overflow-y: auto;
   height: calc(100vh - 5rem);
+  padding: 1rem;
 `;
 
 export const Content = ({ children }: PropsWithChildren<unknown>) => {