+++ packages/stepper/src/StepContent.tsx
@@ -1,10 +1,12 @@
 import { Fragment, Children, ReactElement, cloneElement, PropsWithChildren } from 'react';
 
+/** Define the props interface for StepContent component */
 type StepContentProps = {
   readonly activeStepKey: string | null;
 };
 
 const StepContent = ({ children, activeStepKey }: PropsWithChildren<StepContentProps>): JSX.Element => {
+  /** Find the active step based on the activeStepKey */
   const active = Children.toArray(children).find(
     (child) => (child as ReactElement).props.stepKey === activeStepKey,
   ) as ReactElement;
@@ -13,8 +15,10 @@ const StepContent = ({ children, activeStepKey }: PropsWithChildren<StepContentP
     return <></>;
   }
 
+  /** Clone and map the children of the active step */
   const clonedChildren = Children.map(active.props.children, (child) => cloneElement(child));
 
+  /** Return a Fragment containing the cloned children of the active step */
   return <Fragment>{clonedChildren}</Fragment>;
 };