+++ packages/table/src/components/menu/items/index.ts
@@ -1,2 +1,0 @@
-export * from './ColumnMenuItemSort';
-export * from './ColumnMenuItemFilter';