+++ packages/ripple/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 0.4.1
+
+### Patch Changes
+
+- 4aeee2d4: build(deps): bump @equinor/fusion-wc-ripple from 0.2.21 to 1.1.0
+
 ## 0.4.0
 
 ### Minor Changes